chore: 코드 컨벤션 및 파일/폴더 정리
