--- src/pages/landing-page-legacy/components/BottomGradientOverlay.jsx
@@ -1,7 +1,0 @@
-const BottomGradientOverlay = () => {
-  return (
-    <div className='absolute inset-x-0 bottom-0 h-1/2 bg-gradient-to-b from-transparent to-black' />
-  );
-};
-
-export default BottomGradientOverlay;
