--- src/pages/landing-page-legacy/components/DarkOverlay.jsx
@@ -1,5 +1,0 @@
-const DarkOverlay = () => {
-  return <div className='absolute inset-0 bg-black/30' />;
-};
-
-export default DarkOverlay;
