--- src/pages/landing-page-legacy/components/VinettingOverlay.jsx
@@ -1,7 +1,0 @@
-const VignettingOverlay = () => {
-  return (
-    <div className='pointer-events-none absolute inset-0 z-50 bg-radial from-transparent from-80% to-black/70' />
-  );
-};
-
-export default VignettingOverlay;
