+++ src/pages/landing-page/components/common/DarkLayout.jsx
@@ -1,0 +1,7 @@
+const DarkLayout = () => {
+  return (
+    <div className='absolute inset-0 z-20 bg-gradient-to-b from-black/20 to-black/75' />
+  );
+};
+
+export default DarkLayout;
